+++ src/doc-mirror/index.ts
@@ -60,7 +60,7 @@ export class DocumentModel implements EditableModel {
           }
           if (!options.skipFormat) {
             return formatter.formatPosition(editor, true, {
-              'format-depth': options.formatDepth ? options.formatDepth : 1,
+              'format-depth': options.formatDepth ?? 1,
             });
           }
         }